currencies as double

+++ src/main/java/com/maxtrain/capstone/prs/product/Product.java
@@ -1,16 +1,6 @@
 package com.maxtrain.capstone.prs.product;
 
-import java.math.BigDecimal;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
-import javax.persistence.UniqueConstraint;
+import javax.persistence.*;
 
 import com.maxtrain.capstone.prs.vendor.Vendor;
 
@@ -24,7 +14,7 @@ public class Product {
 	private String partnbr;
 	@Column(length=30, nullable=false)
 	private String name;
-	private BigDecimal price;
+	private double price;
 	@Column(length=30, nullable=false)
 	private String Unit = "Each";
 	@Column(length=255)
@@ -60,11 +50,11 @@ public class Product {
 		this.name = name;
 	}
 
-	public BigDecimal getPrice() {
+	public double getPrice() {
 		return price;
 	}
 
-	public void setPrice(BigDecimal price) {
+	public void setPrice(double price) {
 		this.price = price;
 	}
